Copied definitions of functions resize(), resize2() to .bash_profile.
This ensures that the resize() function is defined and available in
.bash_profile, in case .alias has not been sourced. On openSUSE,
.alias has been sourced when we log in over the serial console, but
that is because their /etc/profile sources $HOME/.bashrc. On other
distros this may not be the case.

I have also left the definitions of the resize funtions in .alias so
they are available in xterms and such, when we have not done a proper
login.