Source .alias at beginning of .bashrc rather than end, and misc formatting cleanups.
Added a section to modify PATH, adding ~/.local/bin and ~/.local/bin/$CPU
directories to the beginning, if they exist. This code uses the path_prepend()
function defined in the .alias file. Thus, we source .alias at the beginning of
.bashrc rather then the end, so its functions are available for use.

Also deleted the line at the very end of the file that was added by 010 Editor,
since we are not using it anymore.