In .bashrc, wrap some commands in tests, to prevent errors on login.
Before sourcing the init-nvm.sh file, we test that it exists on the current
system.

Before adding cabal- and ghcup-related directories to the PATH, we test that
they do in fact exist on the current system.

Before running the command that sets up bash completion for =stack=, we test
that stack is installed.