Added some variable settings to javascript, html and react layers.
In the javascript layer, set ’node-add-modules-path’ to ’t’, which searches for
node_modules/.bin directory by traversing up from the current directory until
reaching the user’s home directory. If it finds node_modules/.bin, it adds it to
the buffer-local variable exec-path. This enables Emacs to find executables of
npm modules installed in the project directory. For instance, it will find and
run ’prettier’ and ’eslint’ installed in the local project.

In the html layer, set html-enable-lsp and css-enable-lsp to ’t’, enabling lsp-
mode for html and css files. We also separately installed the vscode language
servers for html and css using npm. Also set web-fmt-tool to ’prettier.

In the react layer, set web-mode-attr-indent-offset 2.