Moved much code from .zshenv to .zshrc to avoid clobbering on openSUSE.
On openSUSE, /etc/profile gets sourced even in zsh shells, and it resets the
PATH, which clobbered our PATH settings in .zshenv, which are read first.

Also, I learned that settings in .zshenv are read even when #!/bin/zsh scripts
are run, so it seems we should be very careful about what we put in .zshenv.