In .zprofile, set XAUTHORITY to what GDM sets it to so it's set when ssh-ing in.
This allows us to run emacsclient after ssh-ing in to open a new frame
on our machine we ssh-ed in from, as well as all other X client
programs. Without this, emacsclient produced an error because the X
auth key was wrong for our display server. See the files .zprofile and
~/.ssh/rc for more information.