In .spacemacs, removed evil-smartparens as an additional package.
This fixes the issue where in evil visual edit state, 'o' key does not
move to the right.

Spacemacs-evil layer (built-in to spacemacs) already includes the
evil-cleverparens package, which does much the same as
evil-smartparens. The keyboard-layout layer corrects keybindings for
evil-cleverparens, but does not for evil-smartparens. So removing
evil-smartparens lets the already-included evil-cleverparens do its
thing much better.