Rewrote comment of Haskell $PATH section for greater clarity.
This was done after running the GHCup install script on the new Arch
system, which modified .zshrc for us. It sourced the file ~/.ghcup/env
to modify the $PATH for the Haskell Toolchain, but I still want to add
the Haskell directories to the $PATH manually. The way I did it before
is still fine, so no alterations to the actual code were required.
Just rewrote the comment so it is more concise and clear.