In .zshrc & .bashrc, source $HOME/.cargo/env to set up Cargo env vars.
In .bashrc, we already sourced the file $HOME/.cargo/env. In this
commit we just edited the comment a bit for concision and clarity.

In .zshrc, we changed to sourcing the $HOME/.cargo/env file instead of
modifying $PATH directly here. This should be easier to maintain as
upstream adds more settings to that file. This also brings .zshrc and
.bashrc into closer alignment.