In .zprofile, pull in ssh-agent env vars from systemd user environment.
The ssh-agent is now started when a user logs in by systemd user
service, and the systemd user session environment contains the
variables $SSH_AUTH_SOCK and SSH_AGENT_PID. gdm-wayland-session
propagates the systemd user environment into the Wayland session, but
when logging in on a tty we need to do this ourselves.

However, before setting the variables, we check if they are already
set, so as not to overwrite them if they have been set by a forwarded
ssh agent.