Modified tmux.conf for using true-color, and also to improve clarity.
Changed the command to enable true-color in tmux from:

  set-option -sa terminal-overrides ",xterm*:Tc"

to

  set-option -sa terminal-features ",xterm-256color:RGB"

The terminal-features option is a newer addition to tmux, so we assume
it is more appropriate. It is also clearer what it does.

Also edited the comment for the command to set the default-terminal option
to 'tmux-256color' to make it clearer.