Fixed which-key configuration & description in hht-new/configuration.org.
Fixed the problem where you could not page through keys in which-key-mode
if the sequence of the prefix key plus C-h was bound to something. This was the
case for <f1> and <esc>. For <f1> we simply unset the binding of <f1> C-h,
since that was bound to help-for-help which is also reached by <f1> <f1>.
In the case of <esc>, <esc> C-h was bound to mark-defun, and we didn't want to
lose this keybinding, so we moved mark-defun to C-M-y (aka <esc> C-y) and unset
the binding for <esc> C-h.

We also expanded the documentation for which-key and for this issue greatly.