Fix visible scrollbar problem when starting Emacs as server and code cleanup.
We work around the problem in which appearance settings such as theme or hiding
of scrollbars or menu bar do not take effect when Emacs is started as a daemon
although they work fine when Emacs is run from the terminal.

To do this, we defined a function hht/apply-look-and-feel and if we are running
as a daemon, we add this function to the after-make-frame-functions hook so that
it will be called after each frame is created. We apply frame appearance
settings such as hiding the scrollbars in that function.

We also added a function to make frames transparent.

We switched to the Spacemacs dark theme.

We cleaned up and consolidated some code and the documentation in this area.